let's try to get rid of libgui ...
...and "Vector<> have different types" errors

This is an attempt to write a minicap version in kotlin and
use higher level api that should hopefully make it easier
to support many os version (and the future ones). The usage
of SurfaceControl has been inspired by the scrcpy project which
is great.

The implementation is not complete and a few changes will have
to be done on stf.